% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.primary.R
\name{decon.regress1}
\alias{decon.regress1}
\title{Decontaminate metabarcoding data based on a single regression}
\usage{
decon.regress1(data, numb.blanks = 1, taxa = T)
}
\arguments{
\item{data}{A data frame of metabarcoding read data consisting of at least 3 columns in this order: a column of unique OTU names/labels, at least one column of read data from a blank sample (this contains your known contaminant reads), at least one column of read data for an actual sample (each column is a sample, each row is an OTU, and each cell is the number of reads). It can optionally include a final column with taxonomy information. If multiple blanks are included (recommended), they must be in consecutive columns, starting with column 2.}

\item{numb.blanks}{Numeric (default = 1). Specifies the number of blanks included in the data set (if multiple blanks are included, they must be in consecutive columns, starting with column 2).}

\item{taxa}{Logical (T/F). Specifies whether or not the last column contains taxonomic information (default = T)}
}
\value{
A data frame structured the same as the input but with the contaminant reads removed (rows may have been re-ordered).Additionally, all blank columns are condensed into a single mean blank column (the mean of the blanks).
}
\description{
Takes a data frame of metabarcoding reads, identifies the reads that are from contamination, then removes them. It does this based on a single regression line that is used for each sample. It is generally recommended to use decon() rather than decon.regress1()
}
