% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{decon}
\alias{decon}
\title{Decontaminate metabarcoding data based on optimal regressions}
\usage{
decon(data, numb.blanks = 1, taxa = T, runs = 2, low.threshold = 40,
  up.threshold = 400)
}
\arguments{
\item{data}{A data frame of metabarcoding read data consisting of at least 3 columns in this order: a column of unique OTU names/labels, at least one column of read data from a blank sample (this contains your known contaminant reads), at least one column of read data for an actual sample (each column is a sample, each row is an OTU, and each cell is the number of reads). It can optionally include a final column with taxonomy information. If multiple blanks are included (recommended), they must be in consecutive columns, starting with column 2.}

\item{numb.blanks}{Numeric (default = 1). Specifies the number of blanks included in the data set (if multiple blanks are included, they must be in consecutive columns, starting with column 2).}

\item{taxa}{Logical (T/F). Specifies whether or not the last column contains taxonomic information (default = T)}

\item{runs}{Numeric (default = 2). Specifies the number of times that the function should run the decontamination procedure on the data. Based on simulation results, using two runs is best on average.}

\item{low.threshold}{Numeric (default = 40). Selects the lower point for switching between the two decon functions (decon.regress1() and decon.regress2()). It uses decon.regress2 anytime that the estimated overlap is <low.threshold or >up.threshold.}

\item{up.threshold}{Numeric (default = 400). Selects the higher point for switching between the two decon functions (decon.regress1() and decon.regress2()). It uses decon.regress2 anytime that the estimated overlap is <low.threshold or >up.threshold.}
}
\value{
A data frame structured the same as the input but with the contaminant reads removed (rows may have been re-ordered). Additionally, all blank columns are condensed into a single mean blank column (the mean of the blanks).
}
\description{
Takes a data frame of metabacroding reads (structured as a column of OTU IDs, followed by at least one column of reads from blanks, followed by columns of reads from samples, optionally followed by a column of taxonomic information). Ii identifies the reads that are from contamination, then removes them. It estimates the number of overlapping OTUs between the sample and the blank, and it chooses the best equation based on that. Default (recommended) is to use the decon.primary() function anytime that the estimated overlap is <40 or >400, and to use the decon.secondary() function anytime that it is >=40 or <= 400.
}
