% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.thresh.R
\name{remove.thresh}
\alias{remove.thresh}
\title{Set OTUs to zero based on thresholds}
\usage{
remove.thresh(data, numb.ind, taxa = T, thresh = 0.7,
  prop.thresh = 5e-05)
}
\arguments{
\item{data}{A data frame of metabarcoding read data consisting of at least 3 columns in this order: a column of unique OTU names/labels, at least one column of read data from a blank sample (this contains your known contaminant reads), at least one column of read data for an actual sample (each column is a sample, each row is an OTU, and each cell is the number of reads). It can optionally include a final column with taxonomy information. If multiple blanks are included (recommended), they must be in consecutive columns, starting with column 2. Individuals must be ordered by group (e.g., species, populations, etc.)}

\item{numb.ind}{A vector of numbers listing the number of individuals in each user-specified group (e.g., different populations or different species could be treated as different groups). Data must be sorted by these groups beforehand.}

\item{taxa}{Logical (T/F). Specifies whether or not the last column contains taxonomic information (default = T)}

\item{thresh}{Numeric (default = 0.7). A number written as a proportion. This is the threshold at which if that proportion of 0s are present for an OTU within a group, all samples will be set to 0 for that OTU for that group (e.g., if thresh = 0.7, then if, for a particular OTU, 70 percent of samples are 0 within a group, all samples become 0 for that OTU). The threshold always rounds down to calculate the maximum number of zeros that can be present (e.g., if thresh = 0.7 and there are 11 samples, then any OTU with 7 or more 0s will become 0 for all samples in that group). It will not do anything to groups with four or fewer samples. Set to 1 if you do not want to apply this threshold}

\item{prop.thresh}{Numeric (default = 0.00005). A number written as a proportion. This is the threshold at which if the number of reads for a particular OTU are below this proportion, the OTU will be set to zero for all individuals in that group (e.g., if a particular OTU makes up 0.001 percent of all of the reads for a group, then at prop.thresh = 0.00005, that OTU would be set to 0 for all individuals in the group [0.00005 = 0.005 percent]). The proportions are based on all reads for all individuals in a group (including OTUs that were not in the blank). It is necessary to relax this threshold (e.g., 0.0005) for very small data sets (see User Guide section 1.4.4) Set to 0 if you do not want to use this threshold.}
}
\value{
A data frame structured the same as the input but with the unabundant contaminant reads set to zero (rows may have been re-ordered)
}
\description{
Takes a data frame of metabarcoding reads that has gone through remove.cont(), and it finds unabundant OTUs that should be removed and sets them to zero. It only does this for OTUs that amplified in the blank sample. It applies user-specified thresholds to do this, and it does this separately for user-specified groups of individuals. Two threshold arguments are available: thresh and prop.thresh. When both options are utilized, it filters by thresh then by prop.thresh. Both arguments are run independently. Thus, an OTU will be set to zero if either condition is met, rather than if both conditions are met.
}
